refactor: decomposing UserSettings
